benchmark: add x86-64 legs, so each test is measured on two backends
Four series now: xc and Objective-C on this machine, and both again on the
x86-64 Linux host named in build.env. Leaving that variable empty keeps the run
local and reports the skip.

The two legs are built differently, and have to be. xc cross-builds here with
-A x86_64 into a static ELF, so the remote machine needs no toolchain at all.
Objective-C has to compile over there, against libobjc2 and gnustep-base,
because GNUstep Objective-C cannot be cross-built from macOS. Both then time
themselves the same way and report the same checksum, and the runner rejects a
benchmark whose series disagree.

Two portability fixes fell out of running on a second platform:

- CLOCK_MONOTONIC is 6 on Darwin and 1 on Linux, so bench_time.xc selects on
  ARCH_x86_64. One source, both targets.
- int64_t is long on Linux and long long on macOS, so every %lld argument is
  now cast explicitly. Nineteen files; it warned on every Linux build before.

The report draws four bars per optimisation level and its summary table gives
the xc-to-Objective-C ratio per platform, so a row shows at a glance whether a
gap follows the language or the backend.

It also carries a caveat the numbers need: the arm64 Objective-C baseline is
Apple's runtime and Foundation, while the x86-64 one is libobjc2 and
gnustep-base. A ratio compared across the two platforms is also comparing two
Objective-C implementations, which weighs most on the benchmarks dominated by
allocation and dispatch.

First four-series run: nineteen benchmarks, no checksum mismatches. The two
ratio vectors correlate at 0.624, with a mean of 5.75x on arm64 and 4.02x on
x86-64, so part of the gap is shared and part belongs to one backend. array_map
and int_accum are far worse on arm64 than on x86-64, which places them in the
arm64 backend rather than in the IR.

+++ benchmark/src/branch_mix.m
@@ -15,7 +15,7 @@ int main(int argc, char **argv)
             for (uint32_t i = 0; i < N; i++)
                 { if ((a[i] & 1u) == 0u) acc = acc + a[i]; else acc = acc ^ a[i]; }
         int64_t t1 = bench_now_us();
-        printf("%u %lld\n", acc, t1 - t0);
+        printf("%u %lld\n", acc, (long long)(t1 - t0));
     }
     return 0;
     }
