+++ benchmark/src/poly_dispatch.m
@@ -29,7 +29,7 @@ int main(int argc, char **argv)
         for (uint32_t r = 0; r < 20000; r++)
             for (uint32_t i = 0; i < N; i++) acc = [ops[i] apply:acc] + r;
         int64_t t1 = bench_now_us();
-        printf("%u %lld\n", acc, t1 - t0);
+        printf("%u %lld\n", acc, (long long)(t1 - t0));
     }
     return 0;
     }
